+++ src/settings_internal.js
@@ -175,8 +175,6 @@ var EXPECT_MAIN = 1;
 // MODULARIZE, and returned from the factory function.
 var EXPORT_READY_PROMISE = 1;
 
-var USE_LEGACY_DYNCALLS = 0;
-
 // struct_info that is either generated or cached
 var STRUCT_INFO = '';