+++ tools/ports/sdl2_ttf.py
@@ -5,10 +5,10 @@
 
 import os
 
-TAG = 'version_1'
-HASH = '6ce426de0411ba51dd307027c4ef00ff3de4ee396018e524265970039132ab20adb29c2d2e61576c393056374f03fd148dd96f0c4abf8dcee51853dd32f0778f'
+TAG = '703ebc7c66fd' # Latest as of 24 November 2020
+HASH = '362c95f08170a90fe026b264a1053692edea581f5a9908c84aa42d89315cf8c8f74719a80ae45fa16ed7944b12e1b930e5add1d3ad58de04012ed63f535dfb04'
 
-deps = ['freetype', 'sdl2']
+deps = ['freetype', 'sdl2', 'harfbuzz']
 
 
 def needed(settings):
@@ -16,11 +16,10 @@ def needed(settings):
 
 
 def get(ports, settings, shared):
-  ports.fetch_project('sdl2_ttf', 'https://github.com/emscripten-ports/SDL2_ttf/archive/' + TAG + '.zip', 'SDL2_ttf-' + TAG, sha512hash=HASH)
-  libname = 'libSDL2_ttf.a'
+  ports.fetch_project('sdl2_ttf', 'https://hg.libsdl.org/SDL_ttf/archive/' + TAG + '.zip', 'SDL_ttf-' + TAG, sha512hash=HASH)
 
-  def create():
-    src_root = os.path.join(ports.get_dir(), 'sdl2_ttf', 'SDL2_ttf-' + TAG)
+  def create(final):
+    src_root = os.path.join(ports.get_dir(), 'sdl2_ttf', 'SDL_ttf-' + TAG)
     ports.install_headers(src_root, target='SDL2')
 
     srcs = ['SDL_ttf.c']
@@ -31,17 +30,15 @@ def get(ports, settings, shared):
       o = os.path.join(ports.get_build_dir(), 'sdl2_ttf', src + '.o')
       command = [shared.EMCC,
                  '-c', os.path.join(src_root, src),
-                 '-O2', '-s', 'USE_SDL=2', '-s', 'USE_FREETYPE=1', '-o', o, '-w']
+                 '-O2', '-DTTF_USE_HARFBUZZ=1', '-s', 'USE_SDL=2', '-s', 'USE_FREETYPE=1', '-s', 'USE_HARFBUZZ=1', '-o', o, '-w']
       commands.append(command)
       o_s.append(o)
 
     shared.safe_ensure_dirs(os.path.dirname(o_s[0]))
     ports.run_commands(commands)
-    final = os.path.join(ports.get_build_dir(), 'sdl2_ttf', libname)
     ports.create_lib(final, o_s)
-    return final
 
-  return [shared.Cache.get_lib(libname, create, what='port')]
+  return [shared.Cache.get_lib('libSDL2_ttf.a', create, what='port')]
 
 
 def clear(ports, settings, shared):
@@ -51,10 +48,11 @@ def clear(ports, settings, shared):
 def process_dependencies(settings):
   settings.USE_SDL = 2
   settings.USE_FREETYPE = 1
+  settings.USE_HARFBUZZ = 1
 
 
 def process_args(ports):
-  return []
+  return ['-DTTF_USE_HARFBUZZ=1']
 
 
 def show():